Add files via upload
Mr. John Hughes would like you to create a Logistical Regression model and associated ROC/AUC curve for his Happy_Dataset.csv dataset in order to predict if the person is unhappy or happy. To help Mr. Hughes do this, we will create a logistic model to give an outcome when a person is happy or not. We will plot a learning curve graph to see the variance and bias trade off, then we will create a confusion matrix to shows how correctly the model can predict and finalize with the performance with the ROC/AUC curve.